METL-694: further discovered issues with slideIds being in the wrong places.

+++ src/main/scala/com/metl/data/GenericXmlSerializer.scala
@@ -493,7 +493,7 @@ class GenericXmlSerializer(config:ServerConfiguration) extends Serializer with X
       val highlight = getColorByName(bl,"highlight")
       SubmissionBlacklistedPerson(username,highlight)
     }).toList
-    MeTLSubmission(config,m.author,m.timestamp,title,c.slide.toInt,url,imageBytes,blacklist,c.target,c.privacy,c.identity,m.audiences)
+    MeTLSubmission(config,m.author,m.timestamp,title,c.slide,url,imageBytes,blacklist,c.target,c.privacy,c.identity,m.audiences)
   })
   override def fromSubmission(input:MeTLSubmission):NodeSeq = Stopwatch.time("GenericXmlSerializer.fromSubmission",{
     canvasContentToXml("screenshotSubmission",input,List(
